refactor: Move to pman (project manager)
