+++ app/src/main/java/com/iyxan23/sketch/collab/MainActivity.java
@@ -37,6 +37,7 @@ import com.iyxan23.sketch.collab.models.SearchItem;
 import com.iyxan23.sketch.collab.models.SketchwareProject;
 import com.iyxan23.sketch.collab.models.SketchwareProjectChanges;
 import com.iyxan23.sketch.collab.online.BrowseActivity;
+import com.iyxan23.sketch.collab.pman.SketchwareProjectsListActivity;
 
 import org.json.JSONException;
